refactors tests
